pipeline_refactor/pipelines/requirements.txt
absl-py==2.2.2
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
aws-profile-manager==0.7.3
backcall==0.2.0
backports.tarfile==1.2.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
boto3==1.37.3
botocore==1.37.3
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cloudpathlib==0.21.0
cloudpickle==2.2.1
comm==0.2.2
confection==0.1.5
configparser==7.2.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.11
datasets==2.2.1
debugpy==1.8.14
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
docutils==0.21.2
executing==2.2.0
fastjsonschema==2.21.1
fasttext-wheel==0.9.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.41
google-api-core==2.24.2
google-auth==2.40.1
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.2
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
hyperframe==6.1.0
id==1.5.0
idna==3.10
imbalanced-learn==0.12.3
importlib_metadata==8.7.0
ipykernel==6.28.0
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyterlab-pygments==0.1.2
jupyterlab_widgets==3.0.15
keras==3.9.0
keyring==25.6.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libclang==18.1.1
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.9
narwhals==1.38.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.3.7
plotly==5.24.1
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.0
pydantic-settings==2.9.1
pydantic_core==2.33.0
pydeck==0.8.0
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMySQL==1.1.1
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
readme_renderer==44.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.18.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3fs==2025.3.2
s3transfer==0.11.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
sentencepiece==0.2.0
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
srsly==2.5.1
stack-data==0.6.3
streamlit==1.37.1
streamlit-option-menu==0.4.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.49.0
triton==3.3.0
twine==6.1.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.4.0
uv==0.7.3
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xxhash==3.5.0
yarl==1.20.0
zenml==0.82.1
zipp==3.21.0
